Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,6 +3,6 @@ int main()
 {
     int a,b,c;
     scanf("%d%d%d",&a,&b,&c);
-    double avg=(a+b+c)/3;
-    printf("Average: %.2lf",avg);
+    float avg=(a+b+c)/3;
+    printf("Average: %.2f",avg);
 }
